requirements.txt
absl-py==2.1.0
altair==5.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chromedriver-autoinstaller==0.6.4
chromedriver-binary==133.0.6851.0.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cssselect==1.2.0
cycler==0.12.1
distlib==0.3.9
exceptiongroup==1.2.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.55.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lxml==5.3.0
lxml_html_clean==0.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
mediapipe==0.10.18
ml_dtypes==0.5.0
MouseInfo==0.1.3
mpmath==1.3.0
narwhals==1.13.2
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
protobuf==4.25.5
pyarrow==18.0.0
PyAutoGUI==0.9.54
pycparser==2.22
pydeck==0.9.1
pyfiglet==0.7.5
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pynput==1.7.7
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rpds-py==0.21.0
safetensors==0.4.5
scipy==1.13.1
selenium==4.26.1
sentencepiece==0.2.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.6
streamlit==1.39.0
sympy==1.13.1
tenacity==9.0.0
tinysegmenter==0.3
tldextract==5.1.3
tls-client==1.0.1
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1
tornado==6.4.1
tqdm==4.66.6
transformers==4.46.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.1
watchdog==5.0.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.2
wsproto==1.2.0
zipp==3.20.2
